+++ Boolti/Boolti/Sources/Auth/OAuthResponse.swift
@@ -1,0 +1,13 @@
+//
+//  OAuthResponse.swift
+//  Boolti
+//
+//  Created by Miro on 1/23/24.
+//
+
+import Foundation
+
+struct OAuthResponse {
+    let accessToken: String
+    let provider: Provider
+}
