+++ Boolti/Boolti/Sources/Auth/Protocol/SocialLoginAPIServiceType.swift
@@ -7,12 +7,6 @@
 
 import RxSwift
 
-struct OAuthResponse {
-    let accessToken: String
-    let provider: Provider
-}
-
-
 protocol SocialLoginAPIServiceType {
     func authorize(provider: Provider) -> Observable<OAuthResponse>
 }
