+++ Boolti/Boolti/Sources/Auth/Provider.swift
@@ -1,0 +1,13 @@
+//
+//  Provider.swift
+//  Boolti
+//
+//  Created by Miro on 1/23/24.
+//
+
+import Foundation
+
+enum Provider: String {
+    case apple
+    case kakao
+}
