--- Boolti/Boolti/Sources/Network/Foundation/Networking.swift
@@ -1,15 +1,0 @@
-//
-//  Networking.swift
-//  Boolti
-//
-//  Created by Juhyeon Byun on 1/20/24.
-//
-
-import RxSwift
-import RxMoya
-import Moya
-
-protocol Networking {
-    func request(_ api: BaseAPI) -> Single<Response>
-
-}
